requirements.txt
norfair
numpy
opencv-python-headless
kafka-python
confluent-kafka-python
json5
